requirements.txt
pandas
altair
numpy
tabulate
streamlit
google-cloud-aiplatform
python-dotenv
google-cloud-firestore
